+++ lib/src/video_player/method_channel_video_player.dart
@@ -69,6 +69,7 @@ class MethodChannelVideoPlayer extends VideoPlayerPlatform {
           'useCache': dataSource.useCache,
           'maxCacheSize': dataSource.maxCacheSize,
           'maxCacheFileSize': dataSource.maxCacheFileSize,
+          'cacheKey': dataSource.cacheKey,
           'showNotification': dataSource.showNotification,
           'title': dataSource.title,
           'author': dataSource.author,
@@ -77,7 +78,7 @@ class MethodChannelVideoPlayer extends VideoPlayerPlatform {
           'overriddenDuration': dataSource.overriddenDuration?.inMilliseconds,
           'licenseUrl': dataSource.licenseUrl,
           'drmHeaders': dataSource.drmHeaders,
-          'activityName': dataSource.activityName
+          'activityName': dataSource.activityName,
         };
         break;
       case DataSourceType.file:
@@ -278,7 +279,8 @@ class MethodChannelVideoPlayer extends VideoPlayerPlatform {
       'headers': dataSource.headers,
       'maxCacheSize': dataSource.maxCacheSize,
       'maxCacheFileSize': dataSource.maxCacheFileSize,
-      'preCacheSize': preCacheSize
+      'preCacheSize': preCacheSize,
+      'cacheKey': dataSource.cacheKey,
     };
     return _channel.invokeMethod<void>(
       'preCache',