+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added file check for file data source.
 * Fixed issue with black screen for some videos on iOS (by https://github.com/themadmrj)
 * Fixed iOS eventSink issues. (by https://github.com/alextekartik)
+* Added key parameter in BetterPlayerCacheConfiguration to provide way to re-use same video between app sessions.
 
 ## 0.0.68
 * Added support for segmented subtitles.